Move public legacy APIs out of internal
I had moved them into internal by mistake as I moved the code over.
